fixed time string calculation

+++ app/src/main/java/com/andrew/apollo/ui/activities/AudioPlayerActivity.java
@@ -627,7 +627,7 @@ public class AudioPlayerActivity extends AppCompatActivity implements ServiceCon
         // Set the artist name
         mArtistName.setText(MusicUtils.getArtistName());
         // Set the total time
-        mTotalTime.setText(MusicUtils.makeTimeString(this, MusicUtils.duration() / 1000));
+        mTotalTime.setText(MusicUtils.makeTimeString(this, (int) MusicUtils.duration() / 1000));
         // Set the album art
         mImageFetcher.loadCurrentArtwork(mAlbumArt);
         // Set the small artwork
@@ -811,7 +811,7 @@ public class AudioPlayerActivity extends AppCompatActivity implements ServiceCon
     }
 
     private void refreshCurrentTimeText(long pos) {
-        mCurrentTime.setText(MusicUtils.makeTimeString(this, pos / 1000));
+        mCurrentTime.setText(MusicUtils.makeTimeString(this, (int) pos / 1000));
     }
 
     /* Used to update the current time string */
